Change loading of pictures
Pictures aren't supposed to be loaded into memory as they come, or that will
most likely lead to crashes. This implements downsizing (and we'll play with
bitmaps from now on, instead of using ImageView.setImageUri).

Rotation is now possible (and easy!), but I still need to implement a way to
do that from the UI.

(Hopefully) Fixes #13, Fixes #15